add validation function and more strong for security

+++ controller/delete_options.php
@@ -1,4 +1,6 @@
 <?php
-delete_option('custom_field_template_data');
-$options = $this->model->options;
-$message = __('Options deleted.', 'custom-field-template');
+if ( function_exists('current_user_can') && current_user_can('edit_plugins') ) {
+    delete_option('custom_field_template_data');
+    $options = $this->model->options;
+    $message = __('Options deleted.', 'custom-field-template');
+}
